+++ src/game_master.py
@@ -1,0 +1,10 @@
+import random, math, sys
+
+try:
+    pass
+
+except Exception as error:
+    print("Error ", error)
+
+finally:
+    pass
